split entry main and search
