+++ layouts/partials/head.html
@@ -40,6 +40,6 @@
   }
 </script>
 <!-- script -->
-<script src="{{ .Site.BaseURL }}js/bundle.js"></script>
+<script src="{{ .Site.BaseURL }}js/main.bundle.js"></script>
 <script src="{{ .Site.BaseURL }}js/instantpage.min.js" type="module" defer></script>
 {{ hugo.Generator -}}
