+++ layouts/shortcodes/search.html
@@ -7,4 +7,4 @@
   />
 </div>
 <div id="searchResults" class="searchResults"></div>
-<script src="{{ .Site.BaseURL }}js/bundle.js"></script>
+<script src="{{ .Site.BaseURL }}js/search.bundle.js"></script>
