+++ webpack.production.js
@@ -3,8 +3,4 @@ const common = require('./webpack.common.js')
 
 module.exports = merge(common, {
   mode: 'production',
-  output: {
-      path: `${__dirname}/static/js/`,
-      filename: 'bundle.js'
-  }
 })
